+++ packages/moonstone/Slider/SliderTooltip.js
@@ -21,15 +21,25 @@ const SliderTooltipBase = kind({
 
 	propTypes: /** @lends moonstone/Slider.SliderTooltip.prototype */{
 		/**
-		 * Setting to `true` overrides the natural LTR->RTL tooltip side-flipping for locale
-		 * changes. This may be useful if you have a static layout that does not automatically
-		 * reverse when in an RTL language.
+		 * Setting to `true` overrides the natural LTR->RTL tooltip side-flipping for locale changes
+		 * for `vertical` sliders. This may be useful if you have a static layout that does not
+		 * automatically reverse when in an RTL language.
 		 *
 		 * @type {Boolean}
-		 * @default false
 		 * @public
 		 */
 		forceSide: PropTypes.bool,
+
+		/**
+		* When not `vertical`, determines which side of the knob the tooltip appears on.
+		* When `false`, the tooltip will be on the left side, when `true`, the tooltip will
+		* be on the right.
+		*
+		* @type {String}
+		* @default 'rising'
+		* @private
+		*/
+		knobAfterMidpoint: PropTypes.bool,
 
 		/**
 		 * The proportion of progress across the bar. Should be a number between 0 and 1.
@@ -58,13 +68,13 @@ const SliderTooltipBase = kind({
 		 * If `true` the slider will be oriented vertically.
 		 *
 		 * @type {Boolean}
-		 * @default false
 		 * @public
 		 */
 		vertical: PropTypes.bool
 	},
 
 	defaultProps: {
+		knobAfterMidpoint: false,
 		forceSide: false,
 		proportion: 0,
 		side: 'before',
@@ -80,9 +90,9 @@ const SliderTooltipBase = kind({
 
 	computed: {
 		className: ({forceSide, side, vertical, styler}) => styler.append({ignoreLocale: forceSide, vertical, horizontal: !vertical}, side),
-		arrowAnchor: ({proportion, vertical}) => {
+		arrowAnchor: ({knobAfterMidpoint, vertical}) => {
 			if (vertical) return 'middle';
-			return (proportion <= 0.5) ? 'right' : 'left';
+			return knobAfterMidpoint ? 'left' : 'right';
 		},
 		direction: ({forceSide, side, vertical}, context) => {
 			let dir = 'right';
@@ -107,6 +117,7 @@ const SliderTooltipBase = kind({
 	},
 
 	render: ({children, ...rest}) => {
+		delete rest.knobAfterMidpoint;
 		delete rest.forceSide;
 		delete rest.proportion;
 		delete rest.side;