+++ packages/spotlight/CHANGELOG.md
@@ -8,6 +8,8 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Added
 
+- `spotlight/Spottable` ability to restore focus when an initially disabled component becomes enabled
+
 ### Changed
 
 ### Fixed