+++ packages/spotlight/CHANGELOG.md
@@ -8,6 +8,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Added
 
+- `spotlight/SpotlightContainerDecorator` config option `continue5WayHold` to support moving focus to the next spottable element on 5-way hold key.
 - `spotlight/Spottable` ability to restore focus when an initially disabled component becomes enabled
 
 ### Changed