+++ packages/spotlight/CHANGELOG.md
@@ -15,7 +15,7 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
-- `spotlight` to re-focus the last-focused element when window regains focus
+- `spotlight` to correctly restore focus to a spotlight container in another container
 
 ## [1.6.1] - 2017-08-07